print list

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	int i = 1;
+	size_t i = 1;
 
 	if (h == NULL)
 		return (0);
